simplify, use if_nametoindex and if_indextoname
