+++ src/lib/ares__socket.c
@@ -212,14 +212,12 @@ static int configure_socket(ares_socket_t s, struct server_state *server)
     local.sa4.sin_family      = AF_INET;
     local.sa4.sin_addr.s_addr = htonl(channel->local_ip4);
     bindlen                   = sizeof(local.sa4);
-  } else if (server->addr.family == AF_INET6 && ares_strlen(server->ll_iface)) {
+  } else if (server->addr.family == AF_INET6 && ares_strlen(server->ll_scope)) {
     memset(&local.sa6, 0, sizeof(local.sa6));
-    local.sa6.sin6_family = AF_INET6;
+    local.sa6.sin6_family   = AF_INET6;
 #ifdef HAVE_SOCKADDR_IN6_SIN6_SCOPE_ID
     local.sa6.sin6_scope_id = server->ll_scope;
 #endif
-    memcpy(&local.sa6.sin6_addr, &server->ll_addr.addr.addr6,
-           sizeof(local.sa6.sin6_addr));
     bindlen               = sizeof(local.sa6);
   } else if (server->addr.family == AF_INET6 &&
              memcmp(channel->local_ip6, ares_in6addr_any._S6_un._S6_u8,
