+++ src/lib/config-win32.h
@@ -363,6 +363,13 @@
 #  define HAVE_SOCKADDR_IN6_SIN6_SCOPE_ID 1
 #endif
 
+#if defined(_WIN32_WINNT) && (_WIN32_WINNT >= 0x0600)
+/* Define if you have if_nametoindex() */
+#  define HAVE_IF_NAMETOINDEX 1
+/* Define if you have if_indextoname() */
+#  define HAVE_IF_INDEXTONAME 1
+#endif
+
 /* ---------------------------------------------------------------- */
 /*                              Win CE                              */
 /* ---------------------------------------------------------------- */
